feat(routes): /annotation-campaign/new
